Code cleanups in JNAsm.

+++ builder/src/builder/org/jnode/jnasm/assembler/x86/MMX.java
@@ -69,11 +69,11 @@ public class MMX extends AbstractX86Module {
     public boolean emit(String mnemonic, List<Object> operands, int operandSize, Instruction instruction) {
         this.operands = operands;
         this.operandSize = operandSize;
-        Integer key = (Integer) INSTRUCTION_MAP.get(mnemonic);
+        Integer key = INSTRUCTION_MAP.get(mnemonic);
 
         if (key == null) return false;
 
-        switch (key.intValue()) {
+        switch (key) {
             case EMMS_ISN:
                 emitEMMS();
                 break;
@@ -114,7 +114,7 @@ public class MMX extends AbstractX86Module {
                 emitPXOR();
                 break;
             default:
-                throw new Error("Invalid instruction binding " + key.intValue() + " for " + mnemonic);
+                throw new Error("Invalid instruction binding " + key + " for " + mnemonic);
 
         }
         return true;
